Refactor general use files
