+++ src/components/general_use/modal/CustomModal.jsx
@@ -11,8 +11,8 @@ import {
 } from "react-native";
 import Modal from "react-native-modal";
 import { Button, Snackbar, TextInput } from "react-native-paper";
-import AlgoquantApiContext from "../../constants/ApiContext";
-import { THEME } from "../../constants/Theme";
+import AlgoquantApiContext from "../../../constants/ApiContext";
+import { THEME } from "../../../constants/Theme";
 import {
   submitConnectAlpacaModal,
   submitDeleteAccountModal,
@@ -26,8 +26,8 @@ import {
   submitUpdateEmailConfirmNewEmailStep,
   submitUpdateEmailModalNewEmailStep,
   submitUpdatePhoneModal,
-} from "../../helpers/modalSubmitActions";
-import { snackbarCleanUp } from "../../helpers/snackbarCleanup";
+} from "../../../helpers/modalSubmitActions";
+import { snackbarCleanUp } from "../../../helpers/snackbarCleanup";
 
 export default function CustomModal(props) {
   // Get the information for the modal and the functions to clear the information when we close the modal
