+++ src/components/general_use/snackbar/SnackbarContent.jsx
