+++ src/components/main_screens/backtesting/BacktestingScreen.jsx
@@ -14,7 +14,7 @@ import { BACKTEST_FETCH_AMOUNT } from "../../../constants/ApiConstants";
 import AlgoquantApiContext from "../../../constants/ApiContext";
 import { THEME } from "../../../constants/Theme";
 import { backtestHistoryColumns } from "../../../helpers/tableColumns";
-import CustomTable from "../../reusable_components/CustomTable";
+import CustomTable from "../../general_use/table/CustomTable";
 
 export default function BacktestingScreen() {
   const navigation = useNavigation();
