+++ src/components/main_screens/home/create_investor/CreateInvestorStep1Screen.jsx
@@ -13,7 +13,7 @@ import {
 import { Button, Snackbar, TextInput } from "react-native-paper";
 import { THEME } from "../../../../constants/Theme";
 import { snackbarCleanUp } from "../../../../helpers/snackbarCleanup";
-import SnackbarContent from "../../../reusable_components/SnackbarContent";
+import SnackbarContent from "../../../general_use/snackbar/SnackbarContent";
 
 export default function CreateInvestorStep1Screen(props) {
   const navigation = useNavigation();
