+++ src/components/main_screens/home/create_investor/algorithmic/step3/CreateInvestorStockSearch.jsx
@@ -8,7 +8,7 @@ import {
   View,
 } from "react-native";
 import { Searchbar } from "react-native-paper";
-import { THEME, customRNPaperTheme } from "../../constants/Theme";
+import { THEME, customRNPaperTheme } from "../../../../../../constants/Theme";
 
 export default function CreateInvestorStockSearch(props) {
   const {
