+++ src/components/main_screens/profile/options/TradeHistoryScreen.jsx
@@ -11,7 +11,7 @@ import { TRADE_HISTORY_FETCH_AMOUNT } from "../../../../constants/ApiConstants";
 import AlgoquantApiContext from "../../../../constants/ApiContext";
 import { THEME } from "../../../../constants/Theme";
 import { jobHistoryColumns } from "../../../../helpers/tableColumns";
-import CustomTable from "../../../reusable_components/CustomTable";
+import CustomTable from "../../../general_use/table/CustomTable";
 
 export default function TradeHistoryScreen() {
   const [history, setHistory] = useState([]);
