--- src/components/reusable_components/LoadSpinner.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-import { ActivityIndicator } from "react-native";
-
-// Spinner component that is used / shown for loading allowing states to update and render before they are shown.
-const LoadSpinner = () => <ActivityIndicator size="small" color="#3F9F30" />;
-
-export default LoadSpinner;
