+++ src/navigation/StackNavigation.js
@@ -6,9 +6,9 @@ import HomeScreen from "../components/main_screens/home/HomeScreen";
 import CreateInvestorStep1Screen from "../components/main_screens/home/create_investor/CreateInvestorStep1Screen";
 import CreateInvestorAIStep2Screen from "../components/main_screens/home/create_investor/ai/CreateInvestorAIStep2Screen";
 import CreateInvestorAIStep3Screen from "../components/main_screens/home/create_investor/ai/CreateInvestorAIStep3Screen";
-import CreateInvestorAlgorithmicStep2Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep2Screen";
-import CreateInvestorAlgorithmicStep3Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep3Screen";
-import CreateInvestorAlgorithmicStep4Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep4Screen";
+import CreateInvestorAlgorithmicStep2Screen from "../components/main_screens/home/create_investor/algorithmic/step2/CreateInvestorAlgorithmicStep2Screen";
+import CreateInvestorAlgorithmicStep3Screen from "../components/main_screens/home/create_investor/algorithmic/step3/CreateInvestorAlgorithmicStep3Screen";
+import CreateInvestorAlgorithmicStep4Screen from "../components/main_screens/home/create_investor/algorithmic/step4/CreateInvestorAlgorithmicStep4Screen";
 import InvestorScreen from "../components/main_screens/home/invest/investor/InvestorScreen";
 import JobScreen from "../components/main_screens/home/invest/job/JobScreen";
 import ProfileScreen from "../components/main_screens/profile/ProfileScreen";
